refactor: LoggerClient를 Shared Helper로 이전
- LoggerClient 모듈 제거 및 Shared/Helpers/Logger.swift로 이전
- LogLevel, Loggable을 Shared/Types로 이동
- LogInfo 구조체들을 API Client 모듈로 이전
- TCA 의존성 제거하고 정적 메서드로 단순화
- APIClient에서 Logger Helper 직접 호출로 변경
