--- Projects/Client/Logger/Project.swift
@@ -1,15 +1,0 @@
-//
-//  Project.swift
-//
-//  Created by Jiyeon
-//
-
-import ProjectDescription
-import ProjectDescriptionHelpers
-
-let project = Project.buildStaticLibrary(
-  for: Client.Logger,
-  dependencies: [
-    .SPM.TCA
-  ]
-)
